// packages/connect-sdk/src/generated/connections.ts
// AUTO-GENERATED by `npm run generate` in packages/connect-sdk — do not edit by hand.
// Source of truth: Zod schemas in @inboxfm-connect/shared and @inboxfm-connect/core-utils.

export interface ListConnectionsQueryContract {
  cursor?: string
  projectId: string
  scope?: 'PROJECT' | 'PLATFORM'
  pieceName?: string
  displayName?: string
  externalId?: string
  status?: ('ACTIVE' | 'MISSING' | 'ERROR')[]
  limit?: number
}

/**
 * App connection is a connection to an external app.
 */
export interface ConnectionContract {
  id: string
  created: string
  updated: string
  externalId: string
  displayName: string
  type:
    'OAUTH2' | 'PLATFORM_OAUTH2' | 'CLOUD_OAUTH2' | 'SECRET_TEXT' | 'BASIC_AUTH' | 'CUSTOM_AUTH' | 'OIDC' | 'NO_AUTH'
  pieceName: string
  projectIds: string[]
  platformId?: string | null
  scope: 'PROJECT' | 'PLATFORM'
  status: 'ACTIVE' | 'MISSING' | 'ERROR'
  ownerId?: string | null
  owner?: {
    id: string
    email: string
    firstName: string
    status: 'ACTIVE' | 'INACTIVE'
    externalId?: string | null
    platformId?: string | null
    platformRole: 'ADMIN' | 'MEMBER' | 'OPERATOR'
    lastName: string
    created: string
    updated: string
    lastActiveDate?: string | null
    imageUrl?: string | null
  } | null
  metadata?: {
    [k: string]: unknown
  } | null
  pieceVersion: string
  preSelectForNewProjects: boolean
  usingSecretManager: boolean
}

export interface ConnectionsPageContract {
  /**
   * Items: App connection is a connection to an external app.
   */
  data: {
    id: string
    created: string
    updated: string
    externalId: string
    displayName: string
    type:
      'OAUTH2' | 'PLATFORM_OAUTH2' | 'CLOUD_OAUTH2' | 'SECRET_TEXT' | 'BASIC_AUTH' | 'CUSTOM_AUTH' | 'OIDC' | 'NO_AUTH'
    pieceName: string
    projectIds: string[]
    platformId?: string | null
    scope: 'PROJECT' | 'PLATFORM'
    status: 'ACTIVE' | 'MISSING' | 'ERROR'
    ownerId?: string | null
    owner?: {
      id: string
      email: string
      firstName: string
      status: 'ACTIVE' | 'INACTIVE'
      externalId?: string | null
      platformId?: string | null
      platformRole: 'ADMIN' | 'MEMBER' | 'OPERATOR'
      lastName: string
      created: string
      updated: string
      lastActiveDate?: string | null
      imageUrl?: string | null
    } | null
    metadata?: {
      [k: string]: unknown
    } | null
    pieceVersion: string
    preSelectForNewProjects: boolean
    usingSecretManager: boolean
  }[]
  next?: string | null
  previous?: string | null
}